requirements.txt
Flask==3.0.0
numpy==1.24.4
Pillow==9.4.0
Pillow==10.1.0
tensorflow==2.15.0
tensorflow_intel==2.15.0
